+++ app/src/main/AndroidManifest.xml
@@ -21,13 +21,13 @@
         android:theme="@style/AppTheme">
 
         <activity
-            android:name="com.xmartlabs.base.ui.MainActivity"
+            android:name="com.xmartlabs.template.MainActivity"
             android:screenOrientation="portrait"
             />
 
         <activity
             android:label="@string/app_name"
-            android:name="com.xmartlabs.base.ui.StartActivity"
+            android:name="com.xmartlabs.template.StartActivity"
             android:screenOrientation="portrait"
             android:theme="@android:style/Theme.Translucent.NoTitleBar"
             ><!-- Using android:theme="@android:style/Theme.NoDisplay" crashes in API 23. -->
