+++ ui/src/main/java/com/xmartlabs/base/ui/BaseAppCompatActivity.java
@@ -11,6 +11,7 @@ import android.view.inputmethod.InputMethodManager;
 import com.f2prateek.dart.Dart;
 import com.trello.rxlifecycle2.components.RxActivity;
 import com.trello.rxlifecycle2.components.support.RxAppCompatActivity;
+import com.xmartlabs.base.core.Injector;
 
 import bullet.ObjectGraph;
 
@@ -27,7 +28,7 @@ public abstract class BaseAppCompatActivity extends RxAppCompatActivity {
   protected void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
     Dart.inject(this);
-    UIInjector.inject(this);
+    Injector.inject(this);
   }
 
   /**
