--- ui/src/main/java/com/xmartlabs/base/ui/UIComponent.java
@@ -1,15 +1,0 @@
-package com.xmartlabs.base.ui;
-
-import com.xmartlabs.base.ui.module.AndroidModule;
-
-import javax.inject.Singleton;
-
-import dagger.Component;
-
-@Singleton
-@Component(modules = {
-    AndroidModule.class,
-})
-public interface UIComponent {
-
-}
